% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit all the models and return a list object suitable for DFMIP.}
\usage{
fit_models(
  x,
  path_2_case_data = here::here("data-raw/wnv_by_county.csv"),
  case_type = c("neuro", "all"),
  case_variable = "cases"
)
}
\arguments{
\item{x}{tibble or data frame that defines the desired prediction targets. Must have at least a FIPS code and year.}

\item{path_2_case_data}{character file path to the case data}

\item{case_type}{character type of case data}

\item{case_variable}{character name of the case field in the case data}
}
\value{
a list of containing predicted cases for the target in the format expected by \code{DFMIP}
}
\description{
Fit all the models and return a list object suitable for DFMIP.
}
\details{
This function uses the rows in x to define the prediction
targets for the models. There must be at least two fields, the 5 digit fips code
to identify the county as a character variable and a year numeric variable to
identify the year.

The WNV case data cannot be posted publicly as a part of this package. This package
assumes the case data are available in a subdirectory of the working directory
called \code{data-raw}, in a file called \code{wnv_by_county.csv} which has rows
identified by the same fips codes and a year variable. Case counts are assumed
to be in a variable called "cases". There must be at least one row for each
fips code in \code{x}.
}
