% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlwnv.R
\docType{package}
\name{unlwnv}
\alias{unlwnv}
\title{unlwnv: distributed lag function models of wnv annual cases by county}
\description{
Analyzes inter-annual West Nile Virus (WNV) case counts at the county scale using distributed lag models and drought indexes. The package is designed to interface with other analysis packages in the WNV model intercomparison project.
}
