% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_data.R
\name{assemble_data}
\alias{assemble_data}
\title{Assemble WNV data for analysis}
\usage{
assemble_data(
  x,
  path_2_case_data = here::here("data-raw/wnv_by_county.csv"),
  case_type = c("neuro", "all"),
  match_variable = "fips",
  case_variable = "cases",
  measure = NULL,
  nUnits = NULL,
  startUnit = NULL,
  unit = NULL
)
}
\arguments{
\item{x}{a data.frame or tibble with the prediction targets}

\item{path_2_case_data}{character file path to the case data}

\item{case_type}{character type of case data}

\item{match_variable}{character name of the field to match data on}

\item{case_variable}{character name of the case field in the case data}

\item{measure}{character vector of variables to use for lagging}

\item{nUnits}{integer vector of number of units to lag backwards}

\item{startUnit}{integer unit to start lagging from}

\item{unit}{character name of the variable identifying the lag units}
}
\value{
a data.frame ready to estimate distributed lag models
}
\description{
This function pulls together all the different datasets from
package \code{wnvdata} and the local case data. Some basic consistency
checks are performed.
}
